fix: target wizard always showing and PUT target 500 error
- AppRouter read targetData.yearlyTarget instead of targetData.data.yearlyTarget,
  so wizard showed even when target was already configured
- PostgreSQL 42P08 ambiguous parameter type on $2 used as both yearly_target
  (integer) and in $2/12 expression — fixed with explicit cast
- Simplified auth.setup.ts (removed addInitScript that cleared localStorage on reload)

+++ archibald-web-app/backend/src/db/repositories/users.ts
@@ -258,7 +258,7 @@ async function updateUserTarget(
   await pool.query(
     `UPDATE agents.users
      SET yearly_target = $2,
-         monthly_target = $2 / 12,
+         monthly_target = ($2::integer) / 12,
          currency = $3,
          commission_rate = $4,
          bonus_amount = $5,
